requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
pymongo==4.6.1
dnspython==2.4.2